moved slack to service package top level
