+++ service/slack/spec.go
